Move staff/test_models.py into evaluation/test_models.py
